Moved GetPossibleMoves from Grid to Character class to avoid duplication of WrapAroundGridIfRequired method.
