+++ Pacman/Levels/OriginalLayoutLevel.cs
@@ -92,7 +92,6 @@ namespace Pacman
             new Coordinate(8, 5),
             new Coordinate(8, 7),
             new Coordinate(8, 8),
-            new Coordinate(8, 9),
             new Coordinate(8, 10),
             new Coordinate(8, 11),
             new Coordinate(8, 13),
@@ -100,11 +99,6 @@ namespace Pacman
             new Coordinate(8, 16),
             new Coordinate(8, 17),
             new Coordinate(8, 18),
-            new Coordinate(8, 0),
-            new Coordinate(8, 0),
-            new Coordinate(8, 0),
-            new Coordinate(8, 0),
-            new Coordinate(8, 0),
             new Coordinate(9, 7),
             new Coordinate(9, 11),
             new Coordinate(10, 0),
@@ -227,6 +221,10 @@ namespace Pacman
             new Coordinate(7, 16),
             new Coordinate(7, 17),
             new Coordinate(7, 18),
+            new Coordinate(8,9),
+            new Coordinate(9, 8),
+            new Coordinate(9, 9),
+            new Coordinate(9, 10),
             new Coordinate(11, 0),
             new Coordinate(11, 1),
             new Coordinate(11, 2),
