+++ PacmanTests/CharacterTests.cs
@@ -1,0 +1,61 @@
+using System.Collections.Generic;
+using Pacman;
+using Pacman.Input;
+using Pacman.Output;
+using Xunit;
+
+namespace PacmanTests
+{
+    public class CharacterTests
+    {
+        private GridBuilder _gridBuilder = new GridBuilder();
+        
+        [Fact]
+        public void given_CoordinateEqualsOneOne_and_WallsOnEachSideExceptOneTwo_when_GetPossibleMoves_then_return_ListWithOneTwo()
+        {
+            Grid grid = _gridBuilder.GenerateInitialGrid
+            (3,
+                3, 
+                new List<Coordinate>(){new Coordinate(1,0), new Coordinate(2,1), new Coordinate(0,1)},
+                new List<Coordinate>()
+            );
+
+            Character character = new PacmanCharacter(new UserInput(), new ConsoleOutput(), new Coordinate(1, 1));
+
+            List<Coordinate> possibleMoves = character.GetPossibleMoves(new Coordinate(1,1), grid);
+            
+            Assert.Single(possibleMoves);
+            Assert.Equal(1, possibleMoves[0].GetRow());
+            Assert.Equal(2, possibleMoves[0].GetColumn());
+        }
+        
+        [Fact]
+        public void given_CoordinateEqualsOneOne_and_WallsOnTwoSides_when_GetPossibleMoves_then_return_ListWithTwoCoordinates()
+        {
+            Grid grid = _gridBuilder.GenerateInitialGrid
+            (3,
+                3, 
+                new List<Coordinate>(){new Coordinate(1,0), new Coordinate(2,1)},
+                new List<Coordinate>()
+            );
+
+            Character character = new PacmanCharacter(new UserInput(), new ConsoleOutput(), new Coordinate(1, 1));
+
+            List<Coordinate> possibleMoves = character.GetPossibleMoves(new Coordinate(1,1), grid);
+            
+            Assert.Equal(2, possibleMoves.Count);
+        }
+        
+        [Fact]
+        public void given_CoordinateEqualsZeroOne_and_NoWalls_when_GetPossibleMoves_then_return_ListWithFourCoordinates()
+        {
+            Grid grid = _gridBuilder.GenerateEmptyGrid(3, 3);
+
+            Character character = new PacmanCharacter(new UserInput(), new ConsoleOutput(), new Coordinate(1, 1));
+
+            List<Coordinate> possibleMoves = character.GetPossibleMoves(new Coordinate(1,1), grid);
+            
+            Assert.Equal(4, possibleMoves.Count);
+        }
+    }
+}
